api-layer/requirements.txt
azure-functions
azure-identity
python-dotenv
promptflow>=1.8.0
promptflow-azure>=1.8.0
azure-ai-ml>=1.12.0
